+++ analysis/webservice/management/Datasets.py
@@ -18,6 +18,7 @@ import json
 from webservice.NexusHandler import nexus_handler
 from nexustiles.nexustiles import NexusTileService
 from webservice.webmodel import NexusRequestObject, NexusProcessingException
+import logging
 
 from schema import Schema, Or, SchemaError
 from schema import Optional as Opt
@@ -27,6 +28,9 @@ try:
     from yaml import CLoader as Loader
 except ImportError:
     from yaml import Loader
+
+
+logger = logging.getLogger(__name__)
 
 
 CONFIG_SCHEMA = Schema({
@@ -40,10 +44,35 @@ CONFIG_SCHEMA = Schema({
     Opt('aws'): {
         Opt('accessKeyID'): str,
         Opt('secretAccessKey'): str,
+        Opt('profile'): str,
         'public': bool,
         Opt('region'): str
+    },
+    Opt('earthdata'): {
+        Or('endpoint', 'daac'): str,
+        Opt('edl_username'): str,
+        Opt('edl_password'): str,
+    },
+    Opt('mock'): {
+        'accessKeyID': str,
+        'secretAccessKey': str,
     }
 })
+
+
+def validate_config(d):
+    CONFIG_SCHEMA.validate(d)
+
+    # TODO: Can any of these xtra validations be done in the Schema object given the limitations of the schema package?
+
+    if 'aws' in d:
+        assert 'earthdata' not in d
+        if not d['aws']['public']:
+            assert ('accessKeyID' in d['aws'] and 'secretAccessKey' in d['aws']) or 'profile' in d['aws']
+        if 'mock' in d:
+            logger.warning("'mock' present in config with 'aws' credentials config. It will be ignored")
+    elif 'mock' in d:
+        assert 'earthdata' in d
 
 
 class DatasetManagement:
@@ -63,16 +92,8 @@ class DatasetManagement:
             raise NexusProcessingException(reason='Invalid Content-Type header', code=400)
 
         try:
-            CONFIG_SCHEMA.validate(config_dict)
-
-            if 'aws' in config_dict:
-                if not config_dict['aws']['public']:
-                    if 'accessKeyID' not in config_dict['aws'] or 'secretAccessKey' not in config_dict['aws']:
-                        raise NexusProcessingException(
-                            reason='Must provide AWS creds for non-public bucket',
-                            code=400
-                        )
-        except SchemaError as e:
+            validate_config(config_dict)
+        except (SchemaError, AssertionError) as e:
             raise NexusProcessingException(
                 reason=str(e),
                 code=400